create a separate `prompts` module
