+++ chatbot/agents/__init__.py
@@ -1,4 +1,3 @@
-from .prompts import *
 from .react_agent import ReActAgent
 from .reducers import *
 from .router_agent import RouterAgent
