+++ chatbot/prompts/__init__.py
@@ -1,0 +1,3 @@
+from .router_prompts import *
+from .sql_prompts import *
+from .viz_prompts import *
